* commit '8f89bc068419ff5ae50badf87b7be43881faa589':
  Check if intermediate certificate should be passed

+++ php/plugin.smime.php
@@ -619,14 +619,15 @@ class Pluginsmime extends Plugin {
 		// Obtain private certificate
 		$certs = readPrivateCert($this->store, $_SESSION['smime']);
 
-		// Retrieve intermediate CA's for verification.
-		$extracerts = isset($certs['extracerts']) ? implode('', $certs['extracerts']) : '';
-
-		$tmpFile = tempnam(sys_get_temp_dir(), true);
-		file_put_contents($tmpFile, $extracerts);
-
-		$ok = openssl_pkcs7_sign($infile, $outfile, $certs['cert'], array($certs['pkey'], ''), array(), PKCS7_DETACHED, $tmpFile);
-		unlink($tmpFile);
+		// Retrieve intermediate CA's for verification, if avaliable
+		if (isset($certs['extracerts'])) {
+			$tmpFile = tempnam(sys_get_temp_dir(), true);
+			file_put_contents($tmpFile, implode('', $certs['extracerts']));
+			$ok = openssl_pkcs7_sign($infile, $outfile, $certs['cert'], array($certs['pkey'], ''), array(), PKCS7_DETACHED, $tmpFile);
+			unlink($tmpFile);
+		} else {
+			$ok = openssl_pkcs7_sign($infile, $outfile, $certs['cert'], array($certs['pkey'], ''), array(), PKCS7_DETACHED);
+		}
 	}
 
 	/**
